upmed-api/requirements.txt
# Web server
flask
gunicorn
flask_cors

# Code coverage
coverage

# Linting/style checking
pylint
pycodestyle

# External database
firebase-admin
requests

# Reading from .env (.env MUST BE INSIDE upmed-appointment/)
python-dotenv
setuptools

# For basic encryption/decryption
simple-crypt

# JWT
PyJWT


# Twilio API
twilio

